* make compatibility classifer consistently lowercase * undo license reformatting * tab/space fixes * remove commented-out code * add private constructors to static utility classes * add separate default profile * add compat jar to assembly lib dir * hbase.version now defaults to hbase.compat.version * inlcude HBase patch level in compat module names, file names, and classifiers
