+++ phoenix-hbase-compat-2.0.1/src/main/java/org/apache/phoenix/compat/hbase/CompatUtil.java
@@ -1,12 +1,19 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
- * agreements. See the NOTICE file distributed with this work for additional information regarding
- * copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License. You may obtain a
- * copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable
- * law or agreed to in writing, software distributed under the License is distributed on an "AS IS"
- * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License
- * for the specific language governing permissions and limitations under the License.
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package org.apache.phoenix.compat.hbase;
 
@@ -22,6 +29,10 @@ import org.apache.hadoop.hbase.util.Pair;
 
 public class CompatUtil {
 
+    private CompatUtil() {
+        //Not to be instantiated
+    }
+
     public static List<RegionInfo> getMergeRegions(Connection conn, byte[] regionName) throws IOException {
         Pair<RegionInfo, RegionInfo> regionPair =  MetaTableAccessor.getRegionsFromMergeQualifier(conn,regionName);
         List<RegionInfo> regionList = new ArrayList<RegionInfo>(2);
@@ -29,7 +40,7 @@ public class CompatUtil {
         regionList.add(regionPair.getSecond());
         return regionList;
     }
-    
+
     public static int getCellSerializedSize(Cell cell) {
         return org.apache.hadoop.hbase.KeyValueUtil.length(cell);
     }
