+++ phoenix-hbase-compat-2.2.1/pom.xml
@@ -23,17 +23,11 @@
     <artifactId>apache</artifactId>
     <version>21</version>
   </parent>
-<!--   <parent> -->
-<!--     <groupId>org.apache.phoenix</groupId> -->
-<!--     <artifactId>phoenix</artifactId> -->
-<!--     <version>5.1.0-SNAPSHOT</version> -->
-<!--   </parent> -->
-  <groupId>org.apache.phoenix</groupId>
-  <artifactId>phoenix-hbase-compat-2.0</artifactId>
+<groupId>org.apache.phoenix</groupId>
+  <artifactId>phoenix-hbase-compat-2.2.1</artifactId>
   <version>5.1.0-SNAPSHOT</version>
-  
-  <name>Phoenix Hbase 2.0.1 - 2.1 support</name>
-  <description>Phoenix Hbase 2.0.1 - 2.1 support</description>
+  <name>Phoenix Hbase 2.2.1+ support</name>
+  <description>Phoenix Hbase 2.2.1+ support</description>
 
   <build>
     <plugins>
@@ -66,7 +60,7 @@
       <dependency>
         <groupId>org.apache.hbase</groupId>
         <artifactId>hbase-server</artifactId>
-        <version>2.0.1</version>
+        <version>2.2.1</version>
         <scope>compile</scope>
       </dependency>
   </dependencies>
@@ -83,13 +77,4 @@
           </plugin>
       </plugins>
   </reporting>
-  
-      <profiles>
-      <profile>
-      <id>hbase201</id>
-    </profile>
-    <profile>
-      <id>hbase222</id>
-    </profile>
-    </profiles>
 </project>
