+++ phoenix-hbase-compat-2.2.1/src/main/java/org/apache/phoenix/compat/hbase/test/DelegateCell.java
