+++ phoenix-server/pom.xml
@@ -94,7 +94,7 @@
               <goal>shade</goal>
             </goals>
             <configuration>
-                <finalName>phoenix-${project.version}-HBase-${hbase.compat.version}-server</finalName>
+                <finalName>phoenix-${project.version}-${hbase.compat.version.string}-server</finalName>
                 <shadedArtifactAttached>false</shadedArtifactAttached>
                 <promoteTransitiveDependencies>true</promoteTransitiveDependencies>
                 <shadeTestJar>false</shadeTestJar>
